Refactor LevelEntryPoint: add GameController
